+++ admin/tabs/general/config/ConfigPanel.ts
@@ -12,6 +12,7 @@ import {Icon} from '@xh/hoist/icon';
 import {differ} from '../../../differ/Differ';
 import {regroupDialog} from '../../../regroup/RegroupDialog';
 import {ConfigPanelModel} from './ConfigPanelModel';
+import './ConfigPanel.scss';
 
 export const configPanel = hoistCmp.factory({
     model: creates(ConfigPanelModel),
@@ -19,6 +20,7 @@ export const configPanel = hoistCmp.factory({
     render({model}) {
         return fragment(
             restGrid({
+                className: 'xh-config-panel',
                 testId: 'config',
                 extraToolbarItems: () => {
                     return button({